Add files via upload
This project is a billing system where users can purchase items based on data from a provided file (e.g., CSV/JSON). Users provide their name, mobile number, email, and details about the product they want to buy (Product ID and Quantity). The system validates the inputs, calculates the total bill, and updates the stock. It then generates and displays a bill with customer details, product info, and the total amount. The updated stock is saved back into the file for future transactions.

this is the link you can see
http://localhost:8888/notebooks/Inventory%20Mangement%20System.ipynb